test: Use separate test files for each browser
