+++ test/browser/tests/chrome.js
@@ -1,0 +1,4 @@
+const { testBrowser } = require('../browserstack-runner');
+
+const version = '78.0';
+it(`Chrome ${version}`, () => testBrowser('Chrome', version));
