+++ test/browser/tests/edge.js
@@ -1,0 +1,4 @@
+const { testBrowser } = require('../browserstack-runner');
+
+const version = '18.0';
+it(`Edge ${version}`, () => testBrowser('Edge', version));
