+++ test/browser/tests/firefox.js
@@ -1,0 +1,4 @@
+const { testBrowser } = require('../browserstack-runner');
+
+const version = '71.0';
+it(`Firefox ${version}`, () => testBrowser('Firefox', version));
