+++ test/browser/tests/ie.js
@@ -1,0 +1,4 @@
+const { testBrowser } = require('../browserstack-runner');
+
+const version = '11.0';
+it(`IE ${version}`, () => testBrowser('IE', version));
