+++ test/browser/tests/safari.js
@@ -1,0 +1,4 @@
+const { testBrowser } = require('../browserstack-runner');
+
+const version = '12.1';
+it(`Safari ${version}`, () => testBrowser('Safari', version));
